requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.18.0
googlesearch-python==1.3.0
h11==0.14.0
idna==3.10
ipaddress==1.0.23
joblib==1.4.2
motor==3.7.0
numpy==2.2.4
passlib==1.7.4
pydantic==2.11.0
pydantic_core==2.33.0
pymongo==4.11.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-whois==0.9.5
requests==2.32.3
requests-file==2.1.0
scikit-learn==1.5.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.1
threadpoolctl==3.6.0
tldextract==5.1.3
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
